changes related to datacomplianceadmin role

+++ packages/data-product-backend/data-product-lib/src/main/java/com/daimler/data/service/datacompliance/BaseDataComplianceService.java
@@ -32,8 +32,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.persistence.EntityNotFoundException;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -55,7 +53,6 @@ import com.daimler.data.dto.datacompliance.DataComplianceVO;
 import com.daimler.data.service.common.BaseCommonService;
 
 @Service
-@SuppressWarnings(value = "unused")
 public class BaseDataComplianceService extends BaseCommonService<DataComplianceVO, DataComplianceNsql, String>
 		implements DataComplianceService {
 
@@ -110,37 +107,46 @@ public class BaseDataComplianceService extends BaseCommonService<DataComplianceV
 	public ResponseEntity<DataComplianceResponseVO> createDataCompliance(DataComplianceVO requestDataComplianceVO) {
 		DataComplianceResponseVO dataComplianceResponseVO = new DataComplianceResponseVO();
 		try {
-			String uniqueEntityId = requestDataComplianceVO.getEntityId();
-			String uniqueEntityName = requestDataComplianceVO.getEntityName();
-			DataComplianceVO existingDataComplianceVO = super.getByUniqueliteral("entityId", uniqueEntityId);
-			if (existingDataComplianceVO != null && existingDataComplianceVO.getEntityId() != null) {
-				dataComplianceResponseVO.setData(existingDataComplianceVO);
-				List<MessageDescription> messages = new ArrayList<>();
-				MessageDescription message = new MessageDescription();
-				message.setMessage("Record already exists for given entityId.");
-				messages.add(message);
-				dataComplianceResponseVO.setErrors(messages);
-				LOGGER.info("EntityId {} already exists, returning as CONFLICT", uniqueEntityId);
-				return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.CONFLICT);
-			}
-			requestDataComplianceVO.setCreatedBy(this.userStore.getVO());
-			requestDataComplianceVO.setCreatedDate(new Date());
-			requestDataComplianceVO.setId(null);
+			if (verifyUserRoles()) {
+				String uniqueEntityId = requestDataComplianceVO.getEntityId();
+				DataComplianceVO existingDataComplianceVO = super.getByUniqueliteral("entityId", uniqueEntityId);
+				if (existingDataComplianceVO != null && existingDataComplianceVO.getEntityId() != null) {
+					dataComplianceResponseVO.setData(existingDataComplianceVO);
+					List<MessageDescription> messages = new ArrayList<>();
+					MessageDescription message = new MessageDescription();
+					message.setMessage("Record already exists for given entityId.");
+					messages.add(message);
+					dataComplianceResponseVO.setErrors(messages);
+					LOGGER.info("EntityId {} already exists, returning as CONFLICT", uniqueEntityId);
+					return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.CONFLICT);
+				}
+				requestDataComplianceVO.setCreatedBy(this.userStore.getVO());
+				requestDataComplianceVO.setCreatedDate(new Date());
+				requestDataComplianceVO.setId(null);
 
-			DataComplianceVO dataComplianceVO = this.create(requestDataComplianceVO);
-			if (dataComplianceVO != null && dataComplianceVO.getId() != null) {
-				dataComplianceResponseVO.setData(dataComplianceVO);
-				LOGGER.info("Data Compliance entry {} created successfully", uniqueEntityId);
-				return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.CREATED);
+				DataComplianceVO dataComplianceVO = this.create(requestDataComplianceVO);
+				if (dataComplianceVO != null && dataComplianceVO.getId() != null) {
+					dataComplianceResponseVO.setData(dataComplianceVO);
+					LOGGER.info("DataCompliance entry {} created successfully", uniqueEntityId);
+					return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.CREATED);
+				} else {
+					List<MessageDescription> messages = new ArrayList<>();
+					MessageDescription message = new MessageDescription();
+					message.setMessage("Failed to save due to internal error");
+					messages.add(message);
+					dataComplianceResponseVO.setData(requestDataComplianceVO);
+					dataComplianceResponseVO.setErrors(messages);
+					LOGGER.error("Failed to create record with entityId {}", uniqueEntityId);
+					return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.INTERNAL_SERVER_ERROR);
+				}
 			} else {
-				List<MessageDescription> messages = new ArrayList<>();
-				MessageDescription message = new MessageDescription();
-				message.setMessage("Failed to save due to internal error");
-				messages.add(message);
-				dataComplianceResponseVO.setData(requestDataComplianceVO);
-				dataComplianceResponseVO.setErrors(messages);
-				LOGGER.error("Failed to create record with entityId {}", uniqueEntityId);
-				return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.INTERNAL_SERVER_ERROR);
+				List<MessageDescription> notAuthorizedMsgs = new ArrayList<>();
+				MessageDescription notAuthorizedMsg = new MessageDescription();
+				notAuthorizedMsg.setMessage("Not authorized to create dataCompliance entry");
+				notAuthorizedMsgs.add(notAuthorizedMsg);
+				dataComplianceResponseVO.setErrors(notAuthorizedMsgs);
+				LOGGER.debug("DataCompliance entry cannot be created. User not authorized");
+				return new ResponseEntity<>(dataComplianceResponseVO, HttpStatus.FORBIDDEN);
 			}
 		} catch (Exception e) {
 			LOGGER.error("Exception occurred:{} while creating record with entityId {} ", e.getMessage(),
@@ -160,53 +166,62 @@ public class BaseDataComplianceService extends BaseCommonService<DataComplianceV
 	public ResponseEntity<DataComplianceResponseVO> updateDataCompliance(DataComplianceVO requestDataComplianceVO) {
 		DataComplianceResponseVO response = new DataComplianceResponseVO();
 		try {
-			String id = requestDataComplianceVO.getId();
-			DataComplianceVO existingVO = super.getById(id);
-			DataComplianceVO mergedDataComplianceVO = null;
+			if (verifyUserRoles()) {
+				String id = requestDataComplianceVO.getId();
+				DataComplianceVO existingVO = super.getById(id);
+				DataComplianceVO mergedDataComplianceVO = null;
+				if (existingVO != null && existingVO.getId() != null) {
+					String uniqueEntityId = requestDataComplianceVO.getEntityId();
+					DataComplianceVO existingEntityIdVO = super.getByUniqueliteral("entityId", uniqueEntityId);
+					if (existingEntityIdVO != null && existingEntityIdVO.getEntityId() != null
+							&& !existingEntityIdVO.getId().equals(id)) {
+						response.setData(existingEntityIdVO);
+						List<MessageDescription> messages = new ArrayList<>();
+						MessageDescription message = new MessageDescription();
+						message.setMessage("DataCompliance already exists.");
+						messages.add(message);
+						response.setErrors(messages);
+						LOGGER.info("DataCompliance {} already exists, returning as CONFLICT", uniqueEntityId);
+						return new ResponseEntity<>(response, HttpStatus.CONFLICT);
+					}
+					requestDataComplianceVO.setLastModifiedDate(new Date());
+					requestDataComplianceVO.setModifiedBy(this.userStore.getVO());
+					mergedDataComplianceVO = super.create(requestDataComplianceVO);
+					if (mergedDataComplianceVO != null && mergedDataComplianceVO.getId() != null) {
+						response.setData(mergedDataComplianceVO);
+						response.setErrors(null);
+						LOGGER.info("DataCompliance with id {} updated successfully", id);
+						return new ResponseEntity<>(response, HttpStatus.OK);
+					} else {
+						List<MessageDescription> messages = new ArrayList<>();
+						MessageDescription message = new MessageDescription();
+						message.setMessage("Failed to update due to internal error");
+						messages.add(message);
+						response.setData(requestDataComplianceVO);
+						response.setErrors(messages);
+						LOGGER.info("DataCompliance with id {} cannot be edited. Failed with unknown internal error",
+								id);
+						return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+					}
 
-			if (existingVO != null && existingVO.getId() != null) {
-				String uniqueEntityId = requestDataComplianceVO.getEntityId();
-				DataComplianceVO existingEntityIdVO = super.getByUniqueliteral("entityId", uniqueEntityId);
-				if (existingEntityIdVO != null && existingEntityIdVO.getEntityId() != null
-						&& !existingEntityIdVO.getId().equals(id)) {
-					response.setData(existingEntityIdVO);
-					List<MessageDescription> messages = new ArrayList<>();
-					MessageDescription message = new MessageDescription();
-					message.setMessage("DataCompliance already exists.");
-					messages.add(message);
-					response.setErrors(messages);
-					LOGGER.info("DataCompliance {} already exists, returning as CONFLICT", uniqueEntityId);
-					return new ResponseEntity<>(response, HttpStatus.CONFLICT);
-				}
-				requestDataComplianceVO.setLastModifiedDate(new Date());
-				requestDataComplianceVO.setModifiedBy(this.userStore.getVO());
-				mergedDataComplianceVO = super.create(requestDataComplianceVO);
-				if (mergedDataComplianceVO != null && mergedDataComplianceVO.getId() != null) {
-					response.setData(mergedDataComplianceVO);
-					response.setErrors(null);
-					LOGGER.info("DataCompliance with id {} updated successfully", id);
-					return new ResponseEntity<>(response, HttpStatus.OK);
 				} else {
-					List<MessageDescription> messages = new ArrayList<>();
-					MessageDescription message = new MessageDescription();
-					message.setMessage("Failed to update due to internal error");
-					messages.add(message);
-					response.setData(requestDataComplianceVO);
-					response.setErrors(messages);
-					LOGGER.info("DataCompliance with id {} cannot be edited. Failed with unknown internal error", id);
-					return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+					List<MessageDescription> notFoundmessages = new ArrayList<>();
+					MessageDescription notFoundmessage = new MessageDescription();
+					notFoundmessage.setMessage("No DataCompliance found for given id. Update cannot happen");
+					notFoundmessages.add(notFoundmessage);
+					response.setErrors(notFoundmessages);
+					LOGGER.info("No DataCompliance found for given id {} , update cannot happen.", id);
+					return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
 				}
-
 			} else {
-				List<MessageDescription> notFoundmessages = new ArrayList<>();
-				MessageDescription notFoundmessage = new MessageDescription();
-				notFoundmessage.setMessage("No DataCompliance found for given id. Update cannot happen");
-				notFoundmessages.add(notFoundmessage);
-				response.setErrors(notFoundmessages);
-				LOGGER.info("No DataCompliance found for given id {} , update cannot happen.", id);
-				return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+				List<MessageDescription> notAuthorizedMsgs = new ArrayList<>();
+				MessageDescription notAuthorizedMsg = new MessageDescription();
+				notAuthorizedMsg.setMessage("Not authorized to update dataCompliance entry.");
+				notAuthorizedMsgs.add(notAuthorizedMsg);
+				response.setErrors(notAuthorizedMsgs);
+				LOGGER.debug("DataCompliance entry cannot be edited. User not authorized");
+				return new ResponseEntity<>(response, HttpStatus.FORBIDDEN);
 			}
-
 		} catch (Exception e) {
 			LOGGER.error("DataCompliance with id {} cannot be edited. Failed due to internal error {} ",
 					requestDataComplianceVO.getId(), e.getMessage());
@@ -218,43 +233,26 @@ public class BaseDataComplianceService extends BaseCommonService<DataComplianceV
 			response.setErrors(messages);
 			return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
-
-	}
-
-	private boolean canProceedToEdit() {
-		boolean canProceed = false;
-		boolean hasAdminAccess = this.userStore.getUserInfo().hasAdminAccess();
-		if (hasAdminAccess) {
-			canProceed = true;
-		}
-		return canProceed;
 	}
 
 	@Override
 	@Transactional
 	public ResponseEntity<GenericMessage> deleteDataCompliance(String id) {
 		try {
-			DataComplianceVO dataCompliance = this.getById(id);
-//			if (canProceedToEdit()) {
-			this.deleteById(id);
-			GenericMessage successMsg = new GenericMessage();
-			successMsg.setSuccess("success");
-			LOGGER.info("DataCompliance entry with id {} deleted successfully", id);
-			return new ResponseEntity<>(successMsg, HttpStatus.OK);
-//			} else {
-//				MessageDescription notAuthorizedMsg = new MessageDescription();
-//				notAuthorizedMsg.setMessage("Not authorized to delete dataCompliance entry.");
-//				GenericMessage errorMessage = new GenericMessage();
-//				errorMessage.addErrors(notAuthorizedMsg);
-//				LOGGER.debug("DataCompliance enrty with id {} cannot be deleted. User not authorized", id);
-//				return new ResponseEntity<>(errorMessage, HttpStatus.FORBIDDEN);
-//			}
-		} catch (EntityNotFoundException e) {
-			MessageDescription invalidMsg = new MessageDescription("No dataCompliance entry with the given id");
-			GenericMessage errorMessage = new GenericMessage();
-			errorMessage.addErrors(invalidMsg);
-			LOGGER.error("No dataCompliance entry with the given id {} , could not delete.", id);
-			return new ResponseEntity<>(errorMessage, HttpStatus.NOT_FOUND);
+			if (verifyUserRoles()) {
+				this.deleteById(id);
+				GenericMessage successMsg = new GenericMessage();
+				successMsg.setSuccess("success");
+				LOGGER.info("DataCompliance entry with id {} deleted successfully", id);
+				return new ResponseEntity<>(successMsg, HttpStatus.OK);
+			} else {
+				MessageDescription notAuthorizedMsg = new MessageDescription();
+				notAuthorizedMsg.setMessage("Not authorized to delete dataCompliance entry.");
+				GenericMessage errorMessage = new GenericMessage();
+				errorMessage.addErrors(notAuthorizedMsg);
+				LOGGER.debug("DataCompliance entry cannot be deleted. User not authorized");
+				return new ResponseEntity<>(errorMessage, HttpStatus.FORBIDDEN);
+			}
 		} catch (Exception e) {
 			MessageDescription exceptionMsg = new MessageDescription("Failed to delete due to internal error.");
 			GenericMessage errorMessage = new GenericMessage();
